+++ noodles-cram/src/reader/record.rs
@@ -232,8 +232,8 @@ where
         Ok(())
     }
 
-    fn read_read_name(&mut self) -> io::Result<Option<sam::record::ReadName>> {
-        use sam::record::read_name::MISSING;
+    fn read_read_name(&mut self) -> io::Result<Option<sam::record::Name>> {
+        use sam::record::name::MISSING;
 
         let buf = self
             .compression_header
@@ -249,7 +249,7 @@ where
 
         match &buf[..] {
             MISSING => Ok(None),
-            _ => sam::record::ReadName::try_from(buf)
+            _ => sam::record::Name::try_from(buf)
                 .map(Some)
                 .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e)),
         }