+++ noodles-sam/src/record/name.rs
@@ -1,4 +1,4 @@
-//! SAM record read name.
+//! SAM record name.
 
 use std::{
     error, fmt,
@@ -12,21 +12,19 @@ pub const MISSING: &[u8] = b"*";
 // § 1.4 The alignment section: mandatory fields (2020-07-19)
 const MAX_LENGTH: usize = 254;
 
-/// A SAM record read name.
-///
-/// This is also called a query name.
+/// A SAM record name.
 #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct ReadName(pub(crate) Vec<u8>);
+pub struct Name(pub(crate) Vec<u8>);
 
-impl ReadName {
-    /// Creates a read name.
+impl Name {
+    /// Creates a name.
     ///
     /// # Examples
     ///
     /// ```
-    /// use noodles_sam::record::ReadName;
-    /// let read_name = ReadName::try_new("r1")?;
-    /// # Ok::<_, noodles_sam::record::read_name::ParseError>(())
+    /// use noodles_sam::record::Name;
+    /// let name = Name::try_new("r1")?;
+    /// # Ok::<_, noodles_sam::record::name::ParseError>(())
     /// ```
     pub fn try_new<I>(data: I) -> Result<Self, ParseError>
     where
@@ -35,15 +33,15 @@ impl ReadName {
         Self::try_from(data.into())
     }
 
-    /// Returns the length of the read name.
+    /// Returns the length of the name.
     ///
     /// # Examples
     ///
     /// ```
-    /// use noodles_sam::record::ReadName;
-    /// let read_name: ReadName = "r1".parse()?;
-    /// assert_eq!(read_name.len(), 2);
-    /// # Ok::<_, noodles_sam::record::read_name::ParseError>(())
+    /// use noodles_sam::record::Name;
+    /// let name: Name = "r1".parse()?;
+    /// assert_eq!(name.len(), 2);
+    /// # Ok::<_, noodles_sam::record::name::ParseError>(())
     /// ```
     #[allow(clippy::len_without_is_empty)]
     pub fn len(&self) -> usize {
@@ -51,38 +49,38 @@ impl ReadName {
     }
 }
 
-impl crate::alignment::record::Name for ReadName {
+impl crate::alignment::record::Name for Name {
     fn as_bytes(&self) -> &[u8] {
         self.as_ref()
     }
 }
 
-impl crate::alignment::record::Name for &ReadName {
+impl crate::alignment::record::Name for &Name {
     fn as_bytes(&self) -> &[u8] {
         self.as_ref()
     }
 }
 
-impl AsRef<[u8]> for ReadName {
+impl AsRef<[u8]> for Name {
     fn as_ref(&self) -> &[u8] {
         &self.0
     }
 }
 
-impl AsRef<str> for ReadName {
+impl AsRef<str> for Name {
     fn as_ref(&self) -> &str {
         // SAFETY: The internal buffer is limited to ASCII graphic characters (i.e., '!'-'~').
         str::from_utf8(&self.0).unwrap()
     }
 }
 
-impl fmt::Display for ReadName {
+impl fmt::Display for Name {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_str(self.as_ref())
     }
 }
 
-/// An error returned when a raw SAM record read name fails to parse.
+/// An error returned when a raw SAM record name fails to parse.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum ParseError {
     /// The input is empty.
@@ -102,7 +100,7 @@ impl fmt::Display for ParseError {
     }
 }
 
-impl FromStr for ReadName {
+impl FromStr for Name {
     type Err = ParseError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
@@ -110,7 +108,7 @@ impl FromStr for ReadName {
     }
 }
 
-impl TryFrom<Vec<u8>> for ReadName {
+impl TryFrom<Vec<u8>> for Name {
     type Error = ParseError;
 
     fn try_from(buf: Vec<u8>) -> Result<Self, Self::Error> {
@@ -124,9 +122,9 @@ impl TryFrom<Vec<u8>> for ReadName {
     }
 }
 
-impl From<ReadName> for Vec<u8> {
-    fn from(read_name: ReadName) -> Self {
-        read_name.0
+impl From<Name> for Vec<u8> {
+    fn from(name: Name) -> Self {
+        name.0
     }
 }
 
@@ -145,21 +143,21 @@ mod tests {
 
     #[test]
     fn test_fmt() -> Result<(), ParseError> {
-        let read_name: ReadName = "r0".parse()?;
-        assert_eq!(read_name.to_string(), "r0");
+        let name: Name = "r0".parse()?;
+        assert_eq!(name.to_string(), "r0");
         Ok(())
     }
 
     #[test]
     fn test_try_new() {
-        assert_eq!(ReadName::try_new("r0"), Ok(ReadName(b"r0".to_vec())));
+        assert_eq!(Name::try_new("r0"), Ok(Name(b"r0".to_vec())));
 
-        assert_eq!(ReadName::try_new(""), Err(ParseError::Empty));
-        assert_eq!(ReadName::try_new("*"), Err(ParseError::Invalid));
-        assert_eq!(ReadName::try_new("r 0"), Err(ParseError::Invalid));
-        assert_eq!(ReadName::try_new("@r0"), Err(ParseError::Invalid));
+        assert_eq!(Name::try_new(""), Err(ParseError::Empty));
+        assert_eq!(Name::try_new("*"), Err(ParseError::Invalid));
+        assert_eq!(Name::try_new("r 0"), Err(ParseError::Invalid));
+        assert_eq!(Name::try_new("@r0"), Err(ParseError::Invalid));
 
         let s = "n".repeat(MAX_LENGTH + 1);
-        assert_eq!(ReadName::try_new(s), Err(ParseError::Invalid));
+        assert_eq!(Name::try_new(s), Err(ParseError::Invalid));
     }
 }